Move devTools storage connections into seperate package. Allow storage service (sub)class to be passed in
