+++ java/arcs/android/devtools/storage/BUILD
@@ -1,0 +1,30 @@
+load("//third_party/java/arcs/build_defs:build_defs.bzl", "arcs_kt_android_library")
+
+licenses(["notice"])
+
+package(default_visibility = ["//java/arcs:allowed-packages"])
+
+arcs_kt_android_library(
+    name = "devtools",
+    srcs = glob(["*.kt"]),
+    manifest = "//java/arcs/android/common:AndroidManifest.xml",
+    deps = [
+        "//java/arcs/android/crdt",
+        "//java/arcs/android/crdt:crdt_exception_android_proto",
+        "//java/arcs/android/storage",
+        "//java/arcs/android/storage:proxy_message_android_proto",
+        "//java/arcs/android/util",
+        "//java/arcs/core/common",
+        "//java/arcs/core/crdt",
+        "//java/arcs/core/host",
+        "//java/arcs/core/storage",
+        "//java/arcs/core/storage/database",
+        "//java/arcs/core/storage/driver",
+        "//java/arcs/core/storage/util",
+        "//java/arcs/core/util",
+        "//java/arcs/sdk/android/storage",
+        "//third_party/java/androidx/annotation",
+        "//third_party/kotlin/kotlinx_atomicfu",
+        "//third_party/kotlin/kotlinx_coroutines",
+    ],
+)
